+++ deploy-commands.js
@@ -2,16 +2,21 @@ const { REST, Routes } = require('discord.js');
 const config = process.env;
 const token = config.DISCORD_TOKEN;
 const clientId = config.DISCORD_CLIENT_ID;
+const path = require('node:path');
 const fs = require('node:fs');
 
 const commands = [];
-// Grab all the command files from the commands directory you created earlier
-const commandFiles = fs.readdirSync('./commands').filter(file => file.endsWith('.js'));
 
-// Grab the SlashCommandBuilder#toJSON() output of each command's data for deployment
-for (const file of commandFiles) {
-	const command = require(`./commands/${file}`);
-	commands.push(command.data.toJSON());
+
+const commandsPath = path.join(__dirname, 'src/commands');
+const commandsFolders = fs.readdirSync(commandsPath);
+
+for (const folder of commandsFolders) {
+	const files = fs.readdirSync(`${commandsPath}/${folder}`);
+	for (const file of files) {
+		const command = require(`${commandsPath}/${folder}/${file}`);
+		commands.push(command.data.toJSON());
+	}
 }
 
 // Construct and prepare an instance of the REST module
