+++ src/commands/fun/namecolor.js
