+++ src/events/guild/guildMemberRemove.js
